requirements.txt
openai>=1.0.0
azure-ai-openai>=1.0.0
azure-identity
requests
pyyaml
